# openapi/users-openapi.yaml
openapi: "3.0.1"
info:
  title: "Arcadium Users Service"
  description: "The Arcadium Users Service provides a REST API to manage users."
  contact:
    name: "arcadium.dev"
    email: "info@arcadium.dev"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  version: "3.0.1"

paths:
  /v1/user:
    get:
      summary: List returns a list of users.
      description: On successful completion, this returns a list of users.
      operationId: List
      parameters:
        - name: offset
          description: An offset into the total list of users.
          required: false
          in: query
          schema:
            type: string
        - name: limit
          description: An limit to the number of users returned.
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: UsersResponse returns the list of users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "400":
          description: A bad request was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An internal server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create creates a user.
      description: On successful completion, this creates a new user.
      operationId: Create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "400":
          description: A bad request was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The user already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An internal server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/user/{id}:
    get:
      summary: Get returns a user.
      description: On successful completion, this returns a user.
      operationId: Get
      parameters:
        - name: id
          description: The ID of the user.
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: UserResponse returns the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: A bad request was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The user associated with the userID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An internal server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update updates a user.
      description: On successful completion, this updates an existing user.
      operationId: Update
      parameters:
        - name: id
          description: The ID of the user.
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "400":
          description: A bad request was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The user associated with the userID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An internal server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    User:
      type: object
      description: holds a user's information, and is sent in a response.
      required:
        - id
        - login
        - publicKey
        - playerID
        - created
        - updated
      properties:
        id:
          type: string
          x-order: 1
          x-go-name: ID
        login:
          type: string
          x-order: 2
        publicKey:
          type: string
          x-order: 3
        playerID:
          type: string
          x-order: 4
        created:
          type: string
          x-go-type: arcade.Timestamp
          x-go-type-import:
            path: arcadium.dev/arcade
          x-order: 5
        updated:
          type: string
          x-go-type: arcade.Timestamp
          x-go-type-import:
            path: arcadium.dev/arcade
          x-order: 6

    UserResponse:
      type: object
      description: returns a single user in the response.
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'

    UsersResponse:
      type: object
      description: returns a multiple users.
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'

    UserCreateRequest:
      type: object
      description: UserCreateRequest is used to request a user be created.
      required:
        - login
        - publicKey
        - playerID
      properties:
        login:
          type: string
          x-order: 1
        publicKey:
          type: string
          x-order: 2
        playerID:
          type: string
          x-order: 3

    UserUpdateRequest:
      type: object
      description: UserUpdateRequest is used to request a user be updated.
      required:
        - login
        - publicKey
        - playerID
      properties:
        login:
          type: string
          x-order: 1
        publicKey:
          type: string
          x-order: 2
        playerID:
          type: string
          x-order: 3

    ErrorResponse:
      type: object
      description: ErrorResponse provides additional information about problems encounted while performing an operation.
      required:
        - status
        - detail
      properties:
        detail:
          description: Message is a human-readable explanation specific to this occurrence of the problem.
          type: string
        status:
          description: Status is the HTTP status code applicable to this problem.
          type: integer
          format: int